Restructure serial/action-timings
