+++ src/precice/tests/SerialTests.cpp
@@ -1050,223 +1050,6 @@ BOOST_AUTO_TEST_CASE(testConvergenceMeasures3)
   testConvergenceMeasures(configFile, context, expectedIterations);
 }
 
-/**
- * @brief Test to make sure that actions are called in the right order for explicit coupling via RecorderAction
- */
-BOOST_AUTO_TEST_CASE(ActionTimingsExplicit)
-{
-  const std::string configFile = _pathToTests + "recorder-action-explicit.xml";
-  PRECICE_TEST("SolverOne"_on(1_rank), "SolverTwo"_on(1_rank));
-
-  using namespace precice::constants;
-
-  SolverInterface couplingInterface(context.name, configFile, 0, 1);
-
-  int         dimensions = couplingInterface.getDimensions();
-  std::string meshName;
-  std::string writeDataName;
-  std::string readDataName;
-  double      writeValue;
-
-  if (context.isNamed("SolverOne")) {
-    meshName      = "MeshOne";
-    writeDataName = "Forces";
-    readDataName  = "Velocities";
-    writeValue    = 1;
-  } else {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    meshName      = "MeshTwo";
-    writeDataName = "Velocities";
-    readDataName  = "Forces";
-    writeValue    = 2;
-  }
-  int                 meshID      = couplingInterface.getMeshID(meshName);
-  int                 writeDataID = couplingInterface.getDataID(writeDataName, meshID);
-  int                 readDataID  = couplingInterface.getDataID(readDataName, meshID);
-  std::vector<double> vertex(dimensions, 0);
-  int                 vertexID = couplingInterface.setMeshVertex(meshID, vertex.data());
-
-  double dt = -1;
-  BOOST_TEST(action::RecorderAction::records.empty());
-  dt = couplingInterface.initialize();
-  BOOST_TEST(dt == 1.0);
-  if (context.isNamed("SolverOne")) {
-    BOOST_TEST(action::RecorderAction::records.empty());
-  } else {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    BOOST_TEST(action::RecorderAction::records.size() == 2);
-    BOOST_TEST(action::RecorderAction::records.at(0).timing == action::Action::READ_MAPPING_PRIOR);
-    BOOST_TEST(action::RecorderAction::records.at(1).timing == action::Action::READ_MAPPING_POST);
-  }
-  action::RecorderAction::reset();
-  std::vector<double> writeData(dimensions, writeValue);
-  std::vector<double> readData(dimensions, -1);
-  const std::string & cowid = actionWriteInitialData();
-
-  if (couplingInterface.isActionRequired(cowid)) {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    couplingInterface.writeVectorData(writeDataID, vertexID, writeData.data());
-    couplingInterface.markActionFulfilled(cowid);
-  }
-
-  couplingInterface.initializeData();
-  if (context.isNamed("SolverOne")) {
-    BOOST_TEST(action::RecorderAction::records.size() == 2);
-    BOOST_TEST(action::RecorderAction::records.at(0).timing == action::Action::WRITE_MAPPING_PRIOR);
-    BOOST_TEST(action::RecorderAction::records.at(1).timing == action::Action::WRITE_MAPPING_POST);
-  } else {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    BOOST_TEST(action::RecorderAction::records.size() == 4);
-    BOOST_TEST(action::RecorderAction::records.at(0).timing == action::Action::WRITE_MAPPING_PRIOR);
-    BOOST_TEST(action::RecorderAction::records.at(1).timing == action::Action::WRITE_MAPPING_POST);
-    BOOST_TEST(action::RecorderAction::records.at(2).timing == action::Action::READ_MAPPING_PRIOR);
-    BOOST_TEST(action::RecorderAction::records.at(3).timing == action::Action::READ_MAPPING_POST);
-  }
-  action::RecorderAction::reset();
-
-  int iteration = 0;
-
-  while (couplingInterface.isCouplingOngoing()) {
-    couplingInterface.readVectorData(readDataID, vertexID, readData.data());
-    couplingInterface.writeVectorData(writeDataID, vertexID, writeData.data());
-    dt = couplingInterface.advance(dt);
-    BOOST_TEST(couplingInterface.isTimeWindowComplete());
-    iteration++;
-    if (context.isNamed("SolverOne") || iteration < 10) {
-      BOOST_TEST(action::RecorderAction::records.size() == 5);
-      BOOST_TEST(action::RecorderAction::records.at(0).timing == action::Action::WRITE_MAPPING_PRIOR);
-      BOOST_TEST(action::RecorderAction::records.at(1).timing == action::Action::WRITE_MAPPING_POST);
-      BOOST_TEST(action::RecorderAction::records.at(2).timing == action::Action::READ_MAPPING_PRIOR);
-      BOOST_TEST(action::RecorderAction::records.at(3).timing == action::Action::READ_MAPPING_POST);
-      BOOST_TEST(action::RecorderAction::records.at(4).timing == action::Action::ON_TIME_WINDOW_COMPLETE_POST);
-    } else { // SolverTwo only writes in very last iteration, does not read.
-      BOOST_TEST(action::RecorderAction::records.size() == 3);
-      BOOST_TEST(action::RecorderAction::records.at(0).timing == action::Action::WRITE_MAPPING_PRIOR);
-      BOOST_TEST(action::RecorderAction::records.at(1).timing == action::Action::WRITE_MAPPING_POST);
-      BOOST_TEST(action::RecorderAction::records.at(2).timing == action::Action::ON_TIME_WINDOW_COMPLETE_POST);
-    }
-    action::RecorderAction::reset();
-  }
-  couplingInterface.finalize();
-}
-
-/**
- * @brief Test to make sure that actions are called in the right order for implicit coupling via RecorderAction
- */
-BOOST_AUTO_TEST_CASE(ActionTimingsImplicit)
-{
-  PRECICE_TEST("SolverOne"_on(1_rank), "SolverTwo"_on(1_rank));
-  const std::string configFile = _pathToTests + "recorder-action-implicit.xml";
-
-  using namespace precice::constants;
-
-  SolverInterface couplingInterface(context.name, configFile, 0, 1);
-
-  int         dimensions = couplingInterface.getDimensions();
-  std::string meshName;
-  std::string writeDataName;
-  std::string readDataName;
-  std::string writeIterCheckpoint(constants::actionWriteIterationCheckpoint());
-  std::string readIterCheckpoint(constants::actionReadIterationCheckpoint());
-  double      writeValue;
-
-  if (context.isNamed("SolverOne")) {
-    meshName      = "MeshOne";
-    writeDataName = "Forces";
-    readDataName  = "Velocities";
-    writeValue    = 1;
-  } else {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    meshName      = "MeshTwo";
-    writeDataName = "Velocities";
-    readDataName  = "Forces";
-    writeValue    = 2;
-  }
-  int                 meshID      = couplingInterface.getMeshID(meshName);
-  int                 writeDataID = couplingInterface.getDataID(writeDataName, meshID);
-  int                 readDataID  = couplingInterface.getDataID(readDataName, meshID);
-  std::vector<double> vertex(dimensions, 0);
-  int                 vertexID = couplingInterface.setMeshVertex(meshID, vertex.data());
-
-  double dt = -1;
-  BOOST_TEST(action::RecorderAction::records.empty());
-  dt = couplingInterface.initialize();
-  BOOST_TEST(dt == 1.0);
-  if (context.isNamed("SolverOne")) {
-    BOOST_TEST(action::RecorderAction::records.empty());
-  } else {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    BOOST_TEST(action::RecorderAction::records.size() == 2);
-    BOOST_TEST(action::RecorderAction::records.at(0).timing == action::Action::READ_MAPPING_PRIOR);
-    BOOST_TEST(action::RecorderAction::records.at(1).timing == action::Action::READ_MAPPING_POST);
-  }
-  action::RecorderAction::reset();
-  std::vector<double> writeData(dimensions, writeValue);
-  std::vector<double> readData(dimensions, -1);
-  const std::string & cowid = actionWriteInitialData();
-
-  if (couplingInterface.isActionRequired(cowid)) {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    couplingInterface.writeVectorData(writeDataID, vertexID, writeData.data());
-    couplingInterface.markActionFulfilled(cowid);
-  }
-
-  couplingInterface.initializeData();
-  if (context.isNamed("SolverOne")) {
-    BOOST_TEST(action::RecorderAction::records.size() == 2);
-    BOOST_TEST(action::RecorderAction::records.at(0).timing == action::Action::WRITE_MAPPING_PRIOR);
-    BOOST_TEST(action::RecorderAction::records.at(1).timing == action::Action::WRITE_MAPPING_POST);
-  } else {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    BOOST_TEST(action::RecorderAction::records.size() == 4);
-    BOOST_TEST(action::RecorderAction::records.at(0).timing == action::Action::WRITE_MAPPING_PRIOR);
-    BOOST_TEST(action::RecorderAction::records.at(1).timing == action::Action::WRITE_MAPPING_POST);
-    BOOST_TEST(action::RecorderAction::records.at(2).timing == action::Action::READ_MAPPING_PRIOR);
-    BOOST_TEST(action::RecorderAction::records.at(3).timing == action::Action::READ_MAPPING_POST);
-  }
-  action::RecorderAction::reset();
-
-  int iteration = 0;
-
-  while (couplingInterface.isCouplingOngoing()) {
-    couplingInterface.readVectorData(readDataID, vertexID, readData.data());
-    couplingInterface.writeVectorData(writeDataID, vertexID, writeData.data());
-    if (couplingInterface.isActionRequired(writeIterCheckpoint)) {
-      couplingInterface.markActionFulfilled(writeIterCheckpoint);
-    }
-    dt = couplingInterface.advance(dt);
-    if (couplingInterface.isActionRequired(readIterCheckpoint)) {
-      couplingInterface.markActionFulfilled(readIterCheckpoint);
-    }
-    if (couplingInterface.isTimeWindowComplete()) {
-      iteration++;
-    }
-    if (context.isNamed("SolverOne") || iteration < 10) {
-      if (couplingInterface.isTimeWindowComplete()) {
-        BOOST_TEST(action::RecorderAction::records.size() == 5);
-        BOOST_TEST(action::RecorderAction::records.at(4).timing == action::Action::ON_TIME_WINDOW_COMPLETE_POST);
-      } else {
-        BOOST_TEST(action::RecorderAction::records.size() == 4);
-      }
-      BOOST_TEST(action::RecorderAction::records.at(0).timing == action::Action::WRITE_MAPPING_PRIOR);
-      BOOST_TEST(action::RecorderAction::records.at(1).timing == action::Action::WRITE_MAPPING_POST);
-      BOOST_TEST(action::RecorderAction::records.at(2).timing == action::Action::READ_MAPPING_PRIOR);
-      BOOST_TEST(action::RecorderAction::records.at(3).timing == action::Action::READ_MAPPING_POST);
-    } else { // SolverTwo only writes in very last iteration, does not read.
-      if (couplingInterface.isTimeWindowComplete()) {
-        BOOST_TEST(action::RecorderAction::records.size() == 3);
-        BOOST_TEST(action::RecorderAction::records.at(2).timing == action::Action::ON_TIME_WINDOW_COMPLETE_POST);
-      } else {
-        BOOST_TEST(action::RecorderAction::records.size() == 2);
-      }
-      BOOST_TEST(action::RecorderAction::records.at(0).timing == action::Action::WRITE_MAPPING_PRIOR);
-      BOOST_TEST(action::RecorderAction::records.at(1).timing == action::Action::WRITE_MAPPING_POST);
-    }
-    action::RecorderAction::reset();
-  }
-  couplingInterface.finalize();
-}
-
 BOOST_AUTO_TEST_CASE(MultipleFromMappings)
 {
   PRECICE_TEST("A"_on(1_rank), "B"_on(1_rank));
