+++ tests/tests.cmake
@@ -56,6 +56,8 @@ target_sources(testprecice
     tests/serial/access-received-mesh/ExplicitAndMapping.cpp
     tests/serial/access-received-mesh/ExplicitRead.cpp
     tests/serial/access-received-mesh/Implicit.cpp
+    tests/serial/action-timings/ActionTimingsExplicit.cpp
+    tests/serial/action-timings/ActionTimingsImplicit.cpp
     tests/serial/explicit/TestExplicitMPI.cpp
     tests/serial/explicit/TestExplicitMPISingle.cpp
     tests/serial/explicit/TestExplicitSockets.cpp
